+++ src/main/java/racingcar/domain/RaceGame.java
@@ -1,4 +1,6 @@
-package racingcar;
+package racingcar.domain;
+
+import racingcar.ui.ResultView;
 
 import java.util.ArrayList;
 import java.util.List;
