+++ src/main/java/racingcar/util/CarFactory.java
@@ -1,4 +1,6 @@
-package racingcar;
+package racingcar.util;
+
+import racingcar.domain.Car;
 
 import java.util.ArrayList;
 import java.util.List;
