+++ src/main/java/racingcar/util/SpecifiedNumberGenerator.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.util;
 
 public class SpecifiedNumberGenerator implements NumberGenerator {
