+++ src/test/java/racingcar/CarTest.java
@@ -1,9 +1,10 @@
 package racingcar;
 
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+import racingcar.domain.Car;
+import racingcar.util.SpecifiedNumberGenerator;
 
 import static org.assertj.core.api.Assertions.assertThat;
