+++ src/test/java/racingcar/RaceGameTest.java
@@ -2,6 +2,9 @@ package racingcar;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.Car;
+import racingcar.domain.RaceGame;
+import racingcar.util.RandomNumberGenerator;
 
 import java.util.List;
